+++ src/color_matcher.rs
@@ -190,8 +190,8 @@ pub struct FrameMatch {
     pub video_filename: String,
     pub frame_index: u32,
     
-    pub crop_resize: f64,
-    pub crop_pos_x: f64,
-    pub crop_pos_y: f64,
+    pub crop_resize: f32,
+    pub crop_pos_x: f32,
+    pub crop_pos_y: f32,
     pub is_flipped: bool,
 }
